refactor: move utils to projects.utils
