+++ frontend/src/tests/lib/components/project-detail/ProjectTimeline.spec.ts
@@ -6,7 +6,7 @@ import { SnsSwapLifecycle } from "@dfinity/sns";
 import ProjectTimeline from "../../../../lib/components/project-detail/ProjectTimeline.svelte";
 import type { SnsSwapCommitment } from "../../../../lib/types/sns";
 import { secondsToDuration } from "../../../../lib/utils/date.utils";
-import { durationTillSwapDeadline } from "../../../../lib/utils/project.utils";
+import { durationTillSwapDeadline } from "../../../../lib/utils/projects.utils";
 import en from "../../../mocks/i18n.mock";
 import {
   mockSnsFullProject,
